+++ src/fundus/__init__.py
@@ -2,10 +2,9 @@ import pathlib
 import sys
 
 from fundus.publishers import PublisherCollection
-from fundus.scraping.common_crawl import CCNewsCrawler
 from fundus.scraping.filter import Requires
-from fundus.scraping.html import NewsMap, RSSFeed, Sitemap
-from fundus.scraping.pipeline import BaseCrawler, Crawler
+from fundus.scraping.pipeline import BaseCrawler, CCNewsCrawler, Crawler
+from fundus.scraping.url import NewsMap, RSSFeed, Sitemap
 
 __module_path__ = pathlib.Path(__file__).parent
 __development_base_path__ = __module_path__.parents[1]
@@ -16,9 +15,6 @@ __all__ = [
     "CCNewsCrawler",
     "PublisherCollection",
     "Requires",
-    "RSSFeed",
-    "Sitemap",
-    "NewsMap",
 ]
 
 # On a Windows machines, when executing `BaseCrawler.crawl` from our sync API two times,
