+++ src/fundus/publishers/at/__init__.py
@@ -1,5 +1,5 @@
 from fundus.publishers.base_objects import PublisherEnum, PublisherSpec
-from fundus.scraping.html import RSSFeed
+from fundus.scraping.url import RSSFeed
 
 from .orf import OrfParser
