+++ src/fundus/publishers/fr/__init__.py
@@ -1,7 +1,6 @@
 from fundus.publishers.base_objects import PublisherEnum, PublisherSpec
 from fundus.publishers.fr.le_monde import LeMondeParser
-from fundus.scraping.filter import inverse, regex_filter
-from fundus.scraping.html import NewsMap, Sitemap
+from fundus.scraping.url import NewsMap, Sitemap
 
 
 class FR(PublisherEnum):
