+++ src/fundus/publishers/uk/__init__.py
@@ -2,7 +2,7 @@ from datetime import date
 
 from fundus.publishers.base_objects import PublisherEnum, PublisherSpec
 from fundus.scraping.filter import inverse, regex_filter
-from fundus.scraping.html import NewsMap, Sitemap
+from fundus.scraping.url import NewsMap, Sitemap
 
 from .i_news import INewsParser
 from .the_guardian import TheGuardianParser
