--- src/fundus/scraping/common_crawl/__init__.py
@@ -1,3 +1,0 @@
-from .pipeline import CCNewsCrawler
-
-__all__ = ["CCNewsCrawler"]
