+++ src/fundus/scraping/delay.py
@@ -1,0 +1,26 @@
+from __future__ import annotations
+
+import random
+from typing import Protocol, runtime_checkable
+
+
+@runtime_checkable
+class Delay(Protocol):
+    """Protocol to define crawl delays between batches."""
+
+    def __call__(self) -> float:
+        """Yields a float specifying the minimum crawler delay for the current article batch in seconds.
+
+        The effective delay does include crawling execution time between batches,
+        i.e. the effective delay is max(execution_time, delay).
+
+        Examples:
+            >>> import random
+            >>> delay: Delay = lambda: random.random()
+            Will use a random delay in [0, 1) seconds.
+
+        Returns:
+            float: The delay time in seconds.
+
+        """
+        ...
