+++ src/fundus/scraping/scraper.py
@@ -1,55 +1,34 @@
-from typing import AsyncIterator, Literal, Optional
+from typing import Dict, Iterator, List, Literal, Optional, Type
 
 import more_itertools
 
 from fundus.logging import basic_logger
 from fundus.parser import ParserProxy
+from fundus.publishers.base_objects import PublisherEnum
 from fundus.scraping.article import Article
-from fundus.scraping.filter import ExtractionFilter, Requires, URLFilter
-from fundus.scraping.html import FundusSource
+from fundus.scraping.delay import Delay
+from fundus.scraping.filter import ExtractionFilter, URLFilter
+from fundus.scraping.html import CCNewsSource, HTMLSource, WebSource
+from fundus.scraping.url import URLSource
 
 
-class Scraper:
-    def __init__(self, *sources: FundusSource, parser: ParserProxy):
-        self.sources = list(sources)
+class BaseScraper:
+    def __init__(self, *sources: HTMLSource, parser_mapping: Dict[str, ParserProxy]):
+        self.sources = sources
+        self.parser_mapping = parser_mapping
 
-        if not parser:
-            raise ValueError(f"the given parser {type(parser).__name__} is empty")
-
-        self.parser = parser
-
-    async def scrape(
+    def scrape(
         self,
         error_handling: Literal["suppress", "catch", "raise"],
         extraction_filter: Optional[ExtractionFilter] = None,
         url_filter: Optional[URLFilter] = None,
-    ) -> AsyncIterator[Optional[Article]]:
-        # TODO: add docstring; especially explain why returned Article is Optional
-        if isinstance(extraction_filter, Requires):
-            supported_attributes = set(
-                more_itertools.flatten(collection.names for collection in self.parser.attribute_mapping.values())
-            )
-            if missing_attributes := extraction_filter.required_attributes - supported_attributes:
-                if len(missing_attributes) == 1:
-                    basic_logger.warning(
-                        f"The required attribute `{missing_attributes}` "
-                        f"is not supported by {type(self.parser).__name__}. Skipping Scraper"
-                    )
-                else:
-                    basic_logger.warning(
-                        f"The required attributes `{', '.join(missing_attributes)}` "
-                        f"are not supported by {type(self.parser).__name__}. Skipping Scraper"
-                    )
+    ) -> Iterator[Article]:
+        for source in self.sources:
+            for html in source.fetch(url_filter=url_filter):
+                parser = self.parser_mapping[html.source.publisher]
 
-                return
-
-        for html_source in self.sources:
-            async for html in html_source.fetch(url_filter=url_filter):
-                if html is None:
-                    yield None
-                    continue
                 try:
-                    extraction = self.parser(html.crawl_date).parse(html.content, error_handling)
+                    extraction = parser(html.crawl_date).parse(html.content, error_handling)
 
                 except Exception as err:
                     if error_handling == "raise":
@@ -59,16 +38,49 @@ class Scraper:
                         raise err
                     elif error_handling == "catch":
                         yield Article(html=html, exception=err)
-                        continue
                     elif error_handling == "suppress":
                         basic_logger.info(f"Skipped article at '{html.requested_url}' because of: {err!r}")
-                        yield None
                     else:
                         raise ValueError(f"Unknown value '{error_handling}' for parameter <error_handling>'")
 
-                if extraction_filter and extraction_filter(extraction):
-                    basic_logger.debug(f"Skipped article at '{html.requested_url}' because of extraction filter")
-                    yield None
                 else:
-                    article = Article.from_extracted(html=html, extracted=extraction)
-                    yield article
+                    if extraction_filter and extraction_filter(extraction):
+                        basic_logger.debug(f"Skipped article at '{html.requested_url}' because of extraction filter")
+                    else:
+                        article = Article.from_extracted(html=html, extracted=extraction)
+                        yield article
+
+
+class Scraper(BaseScraper):
+    def __init__(
+        self,
+        publisher: PublisherEnum,
+        restrict_sources_to: Optional[List[Type[URLSource]]] = None,
+        delay: Optional[Delay] = None,
+    ):
+        if restrict_sources_to:
+            url_sources = tuple(
+                more_itertools.flatten(publisher.source_mapping[source_type] for source_type in restrict_sources_to)
+            )
+        else:
+            url_sources = tuple(more_itertools.flatten(publisher.source_mapping.values()))
+
+        html_sources = [
+            WebSource(
+                url_source=url_source,
+                publisher=publisher.publisher_name,
+                request_header=publisher.request_header,
+                delay=delay,
+            )
+            for url_source in url_sources
+        ]
+        parser_mapping = {publisher.publisher_name: publisher.parser}
+        super().__init__(*html_sources, parser_mapping=parser_mapping)
+
+
+class CCNewsScraper(BaseScraper):
+    def __init__(self, source: CCNewsSource):
+        parser_mapping: Dict[str, ParserProxy] = {
+            publisher.publisher_name: publisher.parser for publisher in source.publishers
+        }
+        super().__init__(source, parser_mapping=parser_mapping)
